% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcubic.R
\name{xcubic}
\alias{xcubic}
\title{Return cubic of input value}
\usage{
xcubic(.value)
}
\arguments{
\item{.value}{numerical input}
}
\value{
cubed result of input
}
\description{
cubes the input value
}
\examples{
xcubic(2)
}
