% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xsquare.R
\name{xsquare}
\alias{xsquare}
\title{Return square of value}
\usage{
xsquare(.value)
}
\arguments{
\item{.value}{numerical input}
}
\value{
squeared result
}
\description{
Return square of value
}
\examples{
xsquare(4)
xsquare(-2)
}
